// src/types/personalPreferences.ts
export type PersonalPreferences = {
  id: number;
  sexualPreference: "female" | "male";
  commitmentLevel: string;
  personalValues: string[];
  personalBelieves: string[];
  favoriteHolidayDestination: string;
  freeTime: string[];
  musicGenres: string[];
  pet: string;
  userId: number;
  gender: "male" | "female";
  age: number;
  languages: string[];
  height: number;
  bodyType: string;
};